Inline `QueryConstraints` module into `Persistence`
